Ticket #62 and bug 3024749: mesher didn't build with Ogre 1.7.x Took the oportunity to fix a few things, like the old syntax not properly setting the format to "Ogre" (instead of "OgreMesh"), and OgreConverter's help text.
NOTE: This version of the OgreConverter is way inferior to the one in the ogre_branch.
I'm fixing this only to solve all build issues in trunk, and of course similar fixes will be required in the ogre_branch,
but in general I don't expect this module to be very useful.